+++ src/content/index.ts
@@ -1,7 +1,9 @@
 import { init as initHideExplore } from './hide-explore';
 import { init as initHideReels } from './hide-reels';
 import { init as initFilterFeed } from './filter-feed';
+import { init as initVideoControls } from './video-controls';
 
 initHideExplore();
 initHideReels();
 initFilterFeed();
+initVideoControls();